Completely redid searching, moved search logic to be handled by SymbolWindow class, which modifies a SearchContext object and sends it to SearchManager.
Removed ItemManager and CommandManager class, as items and commands are now handled in the SearchContext object in SymbolWindow. Moved much of the results
functionality handling away from Commander class.

Limited search results to 1000 to make search fast, changed itemToCommandMap to be indexed by item type, set Command and Item universe dictionary to be indexed
by the concatenated string of Name+Description, rather than just Name, to allow for items and commands of the same name. 

Changed the name of the Do.cs class to DoMain.cs, because it was interfering with the Do namespace.