Renamed PreferenceChangedEventArgs to PreferencesChangedEventArgs.
Made PreferencesImplementation parameterized by an owner type TOwner instead of just a string representation of the owner. This will make it easier to change the string representation of the owner as it occurs in keypaths if we choose to.

Removed CorePreferences.PreferenceChanged in favor of CorePreferences.ThemeChanged, because ThemeChanged events were all that was being watched for, and this prevents event subscribers from writing down "if (e.Key == "ThemeKey") ...".