LogBase: Fix race condition.
We weren't atomically swapping PendingLogCalls with a new list,
which meant that if a thread was pre-empted between
calls = PendingLogCalls; and
PendingLogCalls = new LinkedList<>();
it was possible for another thread to modify calls under the foreach.
Resulting in sorrow.

Lockless threadsafety is hard, and LogBase still has races in it,
but now one less.